# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data Ingestion & Privacy Intelligence API",
    "version": "1.0.0",
    "description": "Proteccio Discover backend covering Week 1 ingestion, Week 2 discovery/classification/mapping/profiling, and Week 3 governance/compliance intelligence: privacy risk, compliance controls, remediation workflows, dashboards, reporting, RBAC, alerting, and global search.\n\n**Auth:** When `JWT_SECRET` is set, use `POST /api/auth/login` and send `Authorization: Bearer <JWT>`. When only `API_KEY` is set, send `Authorization: Bearer <API_KEY>` or `X-API-Key: <API_KEY>` on `/api/*` except `/health` and documentation routes.\n\n**Docs UI:** `GET /docs` (Swagger UI). **Raw spec:** `GET /openapi.json`."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3000"
        }
      }
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyHeader": []
    }
  ],
  "paths": {
...
    "/api/discovery/scan": {
      "post": {
        "tags": [
          "Discovery"
        ],
        "summary": "Run discovery on record batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DiscoveryScanResult (+ optional classification)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                },
                "example": {
                  "scannedRecords": 1,
                  "findingsPerRecord": [],
                  "summary": {},
                  "trace": {
                    "sourceType": "file",
                    "sourceName": "demo",
                    "entityName": "batch-1"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT from POST /api/auth/login when JWT_SECRET is set; otherwise API_KEY value."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same value as API_KEY when set."
      }
    },
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DiscoveryScanRequest": {
        "type": "object",
        "required": [
          "records"
        ],
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "example": [
              {
                "user": {
                  "email": "a@b.com",
                  "id": 1
                }
              }
            ]
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "database",
              "cloud",
              "file",
              "api"
            ]
          },
          "sourceName": {
            "type": "string"
          },
          "entityName": {
            "type": "string"
          },
          "classify": {
            "type": "boolean"
          },
          "batchSize": {
            "type": "integer",
            "minimum": 1
          },
          "maxDepth": {
            "type": "integer"
          },
          "maxLeavesPerRecord": {
            "type": "integer"
          }
        }
      },
...
    },
...
  }
}